Add function that creates flexible grid
This commit adds a versatile 'createGrid' function, designed to
generate a square grid using the 'desiredGridWidth' variable. The
function  appends the required number of square divs to form a square
grid.

I removed the borders of each square div to give the grid a more
unified appearance, transforming the grid into a cohesive 'screen-like'
interface, rather than a conventional grid structure. An inner
container serves as a border around the entire grid for style.